https://github.com/vertex-protocol/vertex-evm/pull/130#f860076

+++ contracts/common/Constants.sol
@@ -7,8 +7,8 @@ uint32 constant QUOTE_PRODUCT_ID = 0;
 /// @dev Fees account
 uint64 constant FEES_SUBACCOUNT_ID = type(uint64).max;
 
-uint256 constant MINIMUM_LIQUIDITY = 10**3;
+uint128 constant MINIMUM_LIQUIDITY = 10**3;
 
-int256 constant ONE = 10**18;
+int128 constant ONE = 10**18;
 
 uint8 constant MAX_DECIMALS = 18;
