+++ contracts/interfaces/IEndpointGated.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.13;
 
 interface IEndpointGated {
     // this is all that remains lol, everything else is private or a modifier etc.
-    function getOraclePriceX18(uint32 productId) external view returns (int256);
+    function getOraclePriceX18(uint32 productId) external view returns (int128);
 
     function getEndpoint() external view returns (address endpoint);
 }
